fix: rename api/utils to api/_utils for Vercel compatibility
Rename the utils directory to _utils so Vercel ignores it during
serverless function detection. This prevents potential build conflicts
similar to the chat-rooms issue where helper modules were being
processed as API endpoints.

Updated all 18 API files to import from _utils instead of utils.
