+++ api/chat-rooms/_tokens.ts
@@ -4,8 +4,8 @@
 
 import { redis } from "./_redis.js";
 import { CHAT_USERS_PREFIX } from "./_constants.js";
-import { logInfo, logError } from "../utils/logging.js";
-import { isProfaneUsername } from "../utils/validation.js";
+import { logInfo, logError } from "../_utils/logging.js";
+import { isProfaneUsername } from "../_utils/validation.js";
 import {
   generateAuthToken,
   storeToken,
@@ -21,7 +21,7 @@ import {
   setUserPasswordHash,
   TOKEN_GRACE_PERIOD,
   PASSWORD_MIN_LENGTH,
-} from "../utils/auth.js";
+} from "../_utils/auth.js";
 import { createErrorResponse } from "./_helpers.js";
 import type {
   GenerateTokenData,
