+++ api/lyrics.ts
@@ -5,7 +5,7 @@ import {
   getEffectiveOrigin,
   isAllowedOrigin,
   preflightIfNeeded,
-} from "./utils/cors.js";
+} from "./_utils/cors.js";
 
 // Vercel Edge Function configuration
 export const config = {
