+++ api/parse-title.ts
@@ -1,12 +1,12 @@
 import { openai } from "@ai-sdk/openai";
 import { generateObject } from "ai";
 import { z } from "zod";
-import * as RateLimit from "./utils/rate-limit.js";
+import * as RateLimit from "./_utils/rate-limit.js";
 import {
   getEffectiveOrigin,
   isAllowedOrigin,
   preflightIfNeeded,
-} from "./utils/cors.js";
+} from "./_utils/cors.js";
 
 export const config = {
   runtime: "edge",
