+++ api/speech.ts
@@ -1,9 +1,9 @@
 import { experimental_generateSpeech as generateSpeech } from "ai";
 import { openai } from "@ai-sdk/openai";
-import { getEffectiveOrigin, isAllowedOrigin, preflightIfNeeded } from "./utils/cors.js";
-import * as RateLimit from "./utils/rate-limit.js";
+import { getEffectiveOrigin, isAllowedOrigin, preflightIfNeeded } from "./_utils/cors.js";
+import * as RateLimit from "./_utils/rate-limit.js";
 import { Redis } from "@upstash/redis";
-import { validateAuthToken } from "./utils/auth-validate.js";
+import { validateAuthToken } from "./_utils/auth-validate.js";
 
 // --- Default Configuration -----------------------------------------------
