+++ api/youtube-search.ts
@@ -1,10 +1,10 @@
 import { z } from "zod";
-import * as RateLimit from "./utils/rate-limit.js";
+import * as RateLimit from "./_utils/rate-limit.js";
 import {
   getEffectiveOrigin,
   isAllowedOrigin,
   preflightIfNeeded,
-} from "./utils/cors.js";
+} from "./_utils/cors.js";
 
 // Vercel Edge Function configuration
 export const config = {
